+++ dataprovider/foxsports/runner/eventdata/example_test.go
@@ -1,7 +1,9 @@
 package eventdata_test
 
 import (
+	"encoding/json"
 	"fmt"
+	"log"
 
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/foxsports/model"
@@ -32,5 +34,62 @@ func ExampleNBABoxScoreScraper() {
 	boxScoreStats := runner.RunEventsDataScraper(matchups...)
 	for _, statline := range boxScoreStats {
 		fmt.Printf("%#v\n", statline.(model.NBABoxScoreStats))
+	}
+}
+
+// Example for eventdata.MLBBattingBoxScoreScraper
+func ExampleMLBBattingBoxScoreScraper() {
+	// Get matchups
+	matchupRunner := matchup.NewGeneralMatchupRunner(
+		matchup.GeneralMatchupLeague(foxsports.MLB),
+		matchup.GeneralMatchupSegmenter(&foxsports.GeneralSegmenter{Date: "2024-10-30"}),
+	)
+
+	matchups := matchupRunner.GetMatchups()
+
+	// Get boxscore data
+	scraper := eventdata.MLBBattingBoxScoreScraper{}
+	scraper.League = foxsports.MLB
+	runner := eventdata.NewRunner(
+		eventdata.RunnerName("MLB Batting box score stats"),
+		eventdata.RunnerConcurrency(4),
+		eventdata.RunnerScraper(
+			&scraper,
+		),
+	)
+	boxScoreStats := runner.RunEventsDataScraper(matchups...)
+	for _, statline := range boxScoreStats {
+		fmt.Printf("%#v\n", statline.(model.MLBBattingBoxScoreStats))
+	}
+}
+
+// Example for eventdata.MLBPitchingBoxScoreScraper
+func ExampleMLBPitchingBoxScoreScraper() {
+	// Get matchups
+	matchupRunner := matchup.NewGeneralMatchupRunner(
+		matchup.GeneralMatchupLeague(foxsports.MLB),
+		matchup.GeneralMatchupSegmenter(&foxsports.GeneralSegmenter{Date: "2024-10-30"}),
+	)
+
+	matchups := matchupRunner.GetMatchups()
+
+	// Get boxscore data
+	scraper := eventdata.MLBPitchingBoxScoreScraper{}
+	scraper.League = foxsports.MLB
+	runner := eventdata.NewRunner(
+		eventdata.RunnerName("MLB Pitching box score stats"),
+		eventdata.RunnerConcurrency(4),
+		eventdata.RunnerScraper(
+			&scraper,
+		),
+	)
+	boxScoreStats := runner.RunEventsDataScraper(matchups...)
+	// Output each statline as pretty json
+	for _, statline := range boxScoreStats {
+		jsonBytes, err := json.MarshalIndent(statline, "", "  ")
+		if err != nil {
+			log.Fatalf("Error marshaling to JSON: %v\n", err)
+		}
+		fmt.Println(string(jsonBytes))
 	}
 }